requirements.txt
Flask
openpyxl
google-generativeai
markdown
python-docx
gunicorn